team and manager

+++ src/controllers/ProjectController.ts
@@ -4,7 +4,7 @@ import Project from "../models/Project"
 export class ProjectController {
     static createProject = async (req: Request, res: Response) => {
         const project = new Project(req.body)
-
+        project.manager = req.user.id
         try {
             await project.save()
             res.send("Proyecto creado correctamente")
@@ -17,7 +17,11 @@ export class ProjectController {
 
     static getAllprojects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [
+                    {manager: {$in: req.user.id}}
+                ]
+            })
             res.json(projects)
         } catch (error) {
             res.status(500).json({error: 'Hubo un error'})
@@ -34,6 +38,12 @@ export class ProjectController {
                 res.status(404).json({error: error.message})
                 return
             }
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error("Proyecto no encontrado")
+                res.status(404).json({error: error.message})
+                return
+            }
+
             res.json(project)
         } catch (error) {
             res.status(500).json({error: 'Hubo un error'})
@@ -50,6 +60,12 @@ export class ProjectController {
                 res.status(404).json({error: error.message})
                 return
             }
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error("Proyecto no encontrado")
+                res.status(404).json({error: error.message})
+                return
+            }
+
             project.projectName = req.body.projectName
             project.clientName = req.body.clientName
             project.description = req.body.description
@@ -70,6 +86,12 @@ export class ProjectController {
                 res.status(404).json({error: error.message})
                 return
             }
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error("Proyecto no encontrado")
+                res.status(404).json({error: error.message})
+                return
+            }
+
             await project.deleteOne()
             res.send("Proyecto elimnado")
         } catch (error) {
